Clean up PR, add interface
